Split StarcamHelper class to drivers module
